Refactor user handler to user.go
