+++ internal/interfaces/http/user.go
@@ -1,0 +1,49 @@
+package handler
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+
+	"github.com/Financial-Partner/server/internal/contextutil"
+	"github.com/Financial-Partner/server/internal/domain/user"
+)
+
+//go:generate mockgen -source=handler.go -destination=mocks/handler_mock.go -package=mocks
+
+type UserService interface {
+	GetUser(ctx context.Context, email string) (*user.UserEntity, error)
+	GetOrCreateUser(ctx context.Context, email, name string) (*user.UserEntity, error)
+}
+
+// @Summary Get user information
+// @Description Get user information by email from context. If user doesn't exist, a new user will be created.
+// @Tags users
+// @Accept json
+// @Produce json
+// @Security BearerAuth
+// @Param Authorization header string true "Bearer {token}" default "Bearer "
+// @Success 200 {object} user.UserEntity "Successfully retrieved user information"
+// @Failure 400 {string} string "Bad Request - Invalid token format"
+// @Failure 401 {string} string "Unauthorized - Invalid or missing token"
+// @Failure 500 {string} string "Internal server error"
+// @Router /users/me [get]
+func (h *Handler) GetUser(w http.ResponseWriter, r *http.Request) {
+	email := r.Context().Value(contextutil.UserEmailKey)
+	if email == nil {
+		h.log.Errorf("User email not found in context")
+		http.Error(w, "User email not found in context", http.StatusInternalServerError)
+		return
+	}
+
+	logger := h.log.WithField("email", email)
+
+	userEntity, err := h.userService.GetOrCreateUser(r.Context(), email.(string), "")
+	if err != nil {
+		logger.WithError(err).Errorf("Failed to get or create user")
+		http.Error(w, "Failed to retrieve user information", http.StatusInternalServerError)
+		return
+	}
+
+	_ = json.NewEncoder(w).Encode(userEntity)
+}
